+++ src/main/java/com/fiap/techmesa/infrastructure/persistence/repository/OpeningHoursRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fiap.techmesa.infrastructure.persistence.entity.OpeningHoursEntity;
 
-public interface OpeningHoursRepository extends JpaRepository<OpeningHoursEntity, Long>{
+public interface OpeningHoursRepository extends JpaRepository<OpeningHoursEntity, Integer>{
 
 }
