+++ src/main/java/com/fiap/techmesa/infrastructure/persistence/repository/RatingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.fiap.techmesa.infrastructure.persistence.entity.RatingEntity;
 
-public interface RatingRepository extends JpaRepository<RatingEntity, Long>{
+public interface RatingRepository extends JpaRepository<RatingEntity, Integer>{
 
 }
